[cmd/agent] Move some `api` and `app` packages to `internal`
